+++ models/User.js
@@ -39,13 +39,11 @@ const UserSchema = new mongoose.Schema({
     },
     latitude: {
         type: Number,
-        required: [true],
         min: -90,
         max: 90,
     },
     longitude: {
         type: Number,
-        required: [true],
         min: -180,
         max: 180,
     },
